Refactoring yet again. - Added latest tools, targets, and frameworks. - Added coverage display - Added auto deployment
