--- main/Microsoft.Practices.ServiceLocation.PortableClassLibrary.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.21005.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Practices.ServiceLocation.PortableClassLibrary", "Microsoft.Practices.ServiceLocation.PortableClassLibrary\Microsoft.Practices.ServiceLocation.PortableClassLibrary.csproj", "{610653AE-B766-4AD1-A5C6-865246206B8F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceLocation.Tests", "ServiceLocation.Tests\ServiceLocation.Tests.csproj", "{E8106FBA-F925-45CF-9DEB-90806B3E8DBE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{610653AE-B766-4AD1-A5C6-865246206B8F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{610653AE-B766-4AD1-A5C6-865246206B8F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{610653AE-B766-4AD1-A5C6-865246206B8F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{610653AE-B766-4AD1-A5C6-865246206B8F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E8106FBA-F925-45CF-9DEB-90806B3E8DBE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E8106FBA-F925-45CF-9DEB-90806B3E8DBE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E8106FBA-F925-45CF-9DEB-90806B3E8DBE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E8106FBA-F925-45CF-9DEB-90806B3E8DBE}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
