--- main/Microsoft.Practices.ServiceLocation.Silverlight.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 10.00
-# Visual Studio 2008
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Practices.ServiceLocation.Silverlight", "Microsoft.Practices.ServiceLocation.Silverlight\Microsoft.Practices.ServiceLocation.Silverlight.csproj", "{B543BE3E-F1F5-4967-A8AB-B204482C8BA4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B543BE3E-F1F5-4967-A8AB-B204482C8BA4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B543BE3E-F1F5-4967-A8AB-B204482C8BA4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B543BE3E-F1F5-4967-A8AB-B204482C8BA4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B543BE3E-F1F5-4967-A8AB-B204482C8BA4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
