--- main/Microsoft.Practices.ServiceLocation/Properties/AssemblyInfo.Desktop.cs
@@ -1,6 +1,0 @@
-using System;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Security;
-
-[assembly: AllowPartiallyTrustedCallers]
