--- main/ServiceLocation.Tests/Mocks/MockServiceLocator.cs
@@ -1,47 +1,0 @@
-﻿using Microsoft.Practices.ServiceLocation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ServiceLocation.Tests.Mocks
-{
-    class MockServiceLocator : IServiceLocator
-    {
-        public object GetInstance(Type serviceType)
-        {
-            throw new NotImplementedException();
-        }
-
-        public object GetInstance(Type serviceType, string key)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<object> GetAllInstances(Type serviceType)
-        {
-            throw new NotImplementedException();
-        }
-
-        public TService GetInstance<TService>()
-        {
-            throw new NotImplementedException();
-        }
-
-        public TService GetInstance<TService>(string key)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<TService> GetAllInstances<TService>()
-        {
-            throw new NotImplementedException();
-        }
-
-        public object GetService(Type serviceType)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
