+++ package.json
@@ -1,0 +1,5 @@
+{
+  "version": "2.0.0",
+  "name": "Microsoft.Practices.ServiceLocation",
+  "description": "IServiceLocator implementation"
+}
