+++ package.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26730.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Practices.ServiceLocation", "src\Microsoft.Practices.ServiceLocation.csproj", "{CE36BB81-F810-4061-9A59-639C24210D41}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "test\UnitTests.csproj", "{69BC47D2-947A-43BC-A3AB-B810CB2FAE62}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CE36BB81-F810-4061-9A59-639C24210D41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE36BB81-F810-4061-9A59-639C24210D41}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE36BB81-F810-4061-9A59-639C24210D41}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE36BB81-F810-4061-9A59-639C24210D41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{69BC47D2-947A-43BC-A3AB-B810CB2FAE62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{69BC47D2-947A-43BC-A3AB-B810CB2FAE62}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{69BC47D2-947A-43BC-A3AB-B810CB2FAE62}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{69BC47D2-947A-43BC-A3AB-B810CB2FAE62}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {72B4A8E1-7322-442A-8D97-5FAB12C9B0FC}
+	EndGlobalSection
+EndGlobal
