+++ test/Components/ILogger.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved. See License.txt in the project root for license information.
+
+namespace ServiceLocation.Tests.Components
+{
+    public interface ILogger
+    {
+        void Log(string msg);
+    }
+}
