+++ test/Components/SimpleLogger.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved. See License.txt in the project root for license information.
+
+using System;
+
+namespace ServiceLocation.Tests.Components
+{
+    public class SimpleLogger : ILogger
+    {
+        public void Log(string msg)
+        {
+        }
+    }
+}
